requirements.txt
pyautogui
pynput
opencv-python==4.8.1.78
customtkinter
pyaudio
pycolorise
pillow
mss